Update MainWordle.java
There was a bug where last letter label (for word) would not get color 'yellow'. There was a typo for the label name. Now it is fixed.

When enter is hit from the simulated keyboard, it showed a JOptionPane saying "word is invalid". It was due to adding individual if blocks rather than if-else statements in printKeyboardLayout() method. Now it is fixed.